Fix build and allow triggering stats using gamepad.

+++ src/engine_lib/src/render/font_manager.c
@@ -193,8 +193,8 @@ font_manager_cache_glyphs(te_font_manager* manager, unsigned long char_code_firs
         const int gl_format = GL_UNSIGNED_BYTE;
         glBindTexture(GL_TEXTURE_2D, tex_id);
         {
-            glTexImage2D(GL_TEXTURE_2D, 0, GL_LUMINANCE, manager->ft_face->glyph->bitmap.width,
-                         manager->ft_face->glyph->bitmap.rows, 0, GL_LUMINANCE, gl_format,
+            glTexImage2D(GL_TEXTURE_2D, 0, GL_LUMINANCE, (int)manager->ft_face->glyph->bitmap.width,
+                         (int)manager->ft_face->glyph->bitmap.rows, 0, GL_LUMINANCE, gl_format,
                          manager->ft_face->glyph->bitmap.buffer);
 
             glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
